Revert "g4: use vendorsetup script to grab everything we need"
This reverts commit 50a3dfa6d77cd65a7db1022c92762a0d09e40847.

well, that was a very hack'ish way of synching - IMHO  ;)
I prefer using the recommended way whenever it makes sense. its just a
few lines in the local manifest instead of this script and it ensures that
everything is in place as it should when using the repo sync command.

https://github.com/LGgFour/local_manifests/

Change-Id: Ic5a608d0b531f78c771ed07f17e16aa08b1cc29d